refactor(tui): close architectural gaps in hint bar (#309)
Three follow-up gaps from PR #419 review:

1. Move 5 remaining hint consts (RUNNING_VIEW_HINTS, PRESET_SELECT_HINTS,
   GOAL_INPUT_HINTS, LAUNCH_PREVIEW_HINTS, SPAWNING_HINTS, COMPLETE_HINTS)
   from inside 1000-line React component files to lightweight
   src/tui/views/*-hints.ts modules. Matches what advanced-hints.ts and
   panel-hints.ts already do. Decouples hint-map data plane from
   @opentui-ui transitive imports — so pages-nav-acceptance test mock
   can revert to its original minimal 3-export form (workaround
   removed). Removes the corresponding mock hint exports from
   screen-manager.test.ts.

2. Wire useTerminalDimensions into screen-manager so PagesRouter (and
   thus HintBar) responds to actual terminal width rather than the
   hardcoded 100. HintBar's width<40 early-return now fires when the
   real terminal is narrow.

3. Fix 1-5 → 1-4 in RUNNING_VIEW_HINTS to match running-keyboard.ts
   actual key bindings (panels are 1-4; 5 was off-by-one).
