+++ src/tui/screens/agent-detect.tsx
@@ -606,12 +606,3 @@ export const AgentDetect: React.NamedExoticComponent<AgentDetectProps> = React.m
     );
   },
 );
-
-import type { KeyAction } from "../data/hint-map.js";
-
-export const LAUNCH_PREVIEW_HINTS: readonly KeyAction[] = Object.freeze([
-  { key: "Enter", label: "Launch" },
-  { key: "c", label: "CLI" },
-  { key: "e", label: "Edit" },
-  { key: "Esc", label: "Back" },
-]);
