+++ src/tui/screens/complete-view.tsx
@@ -146,10 +146,3 @@ export const CompleteView: React.NamedExoticComponent<CompleteViewProps> = React
     );
   },
 );
-
-import type { KeyAction } from "../data/hint-map.js";
-
-export const COMPLETE_HINTS: readonly KeyAction[] = Object.freeze([
-  { key: "Enter", label: "NewSession" },
-  { key: "q", label: "Quit" },
-]);
