+++ src/tui/screens/preset-select.tsx
@@ -189,11 +189,3 @@ export const PresetSelect: React.NamedExoticComponent<PresetSelectProps> = React
     );
   },
 );
-
-import type { KeyAction } from "../data/hint-map.js";
-
-export const PRESET_SELECT_HINTS: readonly KeyAction[] = Object.freeze([
-  { key: "Enter", label: "Select" },
-  { key: "?", label: "Details" },
-  { key: "q", label: "Quit" },
-]);
