+++ src/tui/screens/running-view.tsx
@@ -1825,14 +1825,3 @@ function renderStatusBar(
     </box>
   );
 }
-
-import type { KeyAction } from "../data/hint-map.js";
-
-export const RUNNING_VIEW_HINTS: readonly KeyAction[] = Object.freeze([
-  { key: ":", label: "Goto" },
-  { key: "/", label: "Filter" },
-  { key: "1-5", label: "Panel" },
-  { key: "Esc", label: "Back" },
-  { key: "?", label: "Help" },
-  { key: "q", label: "Quit" },
-]);
