+++ src/tui/screens/spawn-progress.tsx
@@ -210,9 +210,3 @@ export const SpawnProgress: React.NamedExoticComponent<SpawnProgressProps> = Rea
     );
   },
 );
-
-import type { KeyAction } from "../data/hint-map.js";
-
-export const SPAWNING_HINTS: readonly KeyAction[] = Object.freeze([
-  { key: "Esc", label: "Cancel" },
-]);
