+++ src/tui/views/spawning-hints.ts
@@ -1,0 +1,6 @@
+/** Hints for the spawning screen (#309). */
+import type { KeyAction } from "../data/hint-map.js";
+
+export const SPAWNING_HINTS: readonly KeyAction[] = Object.freeze([
+  { key: "Esc", label: "Cancel" },
+]);
